fix(db): use bigint for timestamps to avoid PG integer out of range

+++ src/domain/entities/PlaybackState.ts
@@ -31,6 +31,6 @@ export class PlaybackState {
   @Column({ default: false })
   shuffle!: boolean;
 
-  @Column({ type: 'integer', default: Date.now() })
+  @Column({ type: 'bigint', default: Date.now() })
   updatedAt!: number;
 }
